feature: Introduce run_on_init flag
Run the rule's commands once before when flag with "run_on_init".

e.g.:

- name: run my test
  run: 'make tests'
  change: 'src/**'
  run_on_init: true

This would run the `make tests` once on start, even if nothing changed,
and then start the watcher.

The flag is set to false by default, if the key is not present on the
yaml or if values can't be assumed.